% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_site}
\alias{build_site}
\title{Build a complete pkgdown website offline}
\usage{
build_site(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}.}
}
\description{
Build a complete pkgdown website offline
}
