% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_from_cache}
\alias{copy_from_cache}
\title{Copy deps from cache to somewhere (for example, pkgdown.offline in dev mode)}
\usage{
copy_from_cache(version, destdir)
}
\description{
Copy deps from cache to somewhere (for example, pkgdown.offline in dev mode)
}
