% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_site}
\alias{init_site}
\title{Initialize site infrastructure offline}
\usage{
init_site(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[pkgdown:init_site]{pkgdown::init_site()}}.}
}
\value{
Invisible \code{NULL}, called for side effect of initializing site resources.
}
\description{
Sets up the necessary infrastructure for building a pkgdown site offline
by copying locally cached dependencies to the pkgdown cache directory.
}
\examples{
\dontrun{
pkgdown.offline::init_site(override = list(destination = tempdir()))
}
}
