% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_cache_dev}
\alias{path_cache_dev}
\title{Path to cache directory in pkgdown.offline (source state)}
\usage{
path_cache_dev()
}
\value{
Character string with the path to the cache directory.
}
\description{
Returns the path to the cache directory when the package is in
development mode (not installed).
}
\examples{
# Get the path to the cache in development mode
pkgdown.offline::path_cache_dev()
}
