% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_cache_installed}
\alias{path_cache_installed}
\title{Path to cache directory in pkgdown.offline (installed state)}
\usage{
path_cache_installed()
}
\value{
Character string with the path to the cache directory.
}
\description{
Returns the path to the cache directory when the package is installed.
}
\examples{
# Get the path to the cache in the installed package
pkgdown.offline::path_cache_installed()
}
