% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_cache}
\alias{update_cache}
\title{Cache pkgdown dependencies}
\usage{
update_cache(version, destdir = tempdir())
}
\arguments{
\item{version}{pkgdown version.}

\item{destdir}{Where to put the cached dependencies.}
}
\description{
Cache pkgdown dependencies for each pkgdown version
(>= 2.1.0) on CRAN. Running This requires internet connection.
}
\note{
To update the cache in pkgdown.offline,
run \code{update_cache(..., destdir = path_pkg_cache())}.
}
\examples{
update_cache("2.1.0")
update_cache("2.1.1")
update_cache("2.1.2")
}
