% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_cache_installed}
\alias{path_cache_installed}
\title{Path to cache directory in pkgdown.offline (installed state)}
\usage{
path_cache_installed()
}
\description{
Path to cache directory in pkgdown.offline (installed state)
}
