% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.offline-package.R
\docType{package}
\name{pkgdown.offline-package}
\alias{pkgdown.offline}
\alias{pkgdown.offline-package}
\title{pkgdown.offline: Build 'pkgdown' Websites Offline}
\description{
Build 'pkgdown' websites without an internet connection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/pkgdown.offline/}
  \item \url{https://github.com/nanxstats/pkgdown.offline}
  \item Report bugs at \url{https://github.com/nanxstats/pkgdown.offline/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Authors:
\itemize{
  \item John Blischak (\href{https://orcid.org/0000-0003-2634-9879}{ORCID})
}

}
\keyword{internal}
