% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.offline-package.R
\docType{package}
\name{pkgdown.offline-package}
\alias{pkgdown.offline}
\alias{pkgdown.offline-package}
\title{pkgdown.offline: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
