% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.offline-package.R
\docType{package}
\name{pkgdown.offline-package}
\alias{pkgdown.offline}
\alias{pkgdown.offline-package}
\title{pkgdown.offline: Build 'pkgdown' Websites Offline}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Build 'pkgdown' websites without an internet connection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/pkgdown.offline/}
  \item \url{https://github.com/nanxstats/pkgdown.offline}
  \item Report bugs at \url{https://github.com/nanxstats/pkgdown.offline/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item John Blischak (\href{https://orcid.org/0000-0003-2634-9879}{ORCID})
}

Other contributors:
\itemize{
  \item Algolia, Inc. and other contributors (autocomplete.js library) [contributor, copyright holder]
  \item Aidan Feldman (bootstrap-toc library) [contributor, copyright holder]
  \item Zeno Rocha (clipboard.js library) [contributor, copyright holder]
  \item Nick Williams (headroom.js library) [contributor, copyright holder]
  \item Julian Kühnel (mark.js library) [contributor, copyright holder]
  \item Kiro Risk (Fuse.js library) [contributor, copyright holder]
  \item Khan Academy and other contributors (KaTeX library) [contributor, copyright holder]
  \item The MathJax Consortium (MathJax library) [contributor, copyright holder]
}

}
\keyword{internal}
